feat(RF2): Soporte de unidades de medida en productos y cambio de valor de la variable ID
- Agregado enum UnidadMedida con valores UNIDAD, KG, GR, LT, ML
- Actualizado la entidad Producto para incluir el campo unidadMedida mapeado con @Enumerated(EnumType.STRING)
- Modificado ProductoController en PUT /api/productos/{id} para permitir actualizar unidadMedida
- Modificado el tipo Long por tipo Integer en el campo ID

+++ src/main/java/com/irso/innovasys/service/ProductoService.java
@@ -25,12 +25,12 @@ public class ProductoService {
     }
 
     // Buscar producto por ID
-    public Optional<Producto> obtenerProductoPorId(Long id) {
+    public Optional<Producto> obtenerProductoPorId(Integer id) {
         return productoRepository.findById(id);
     }
 
     // Eliminar producto por ID
-    public void eliminarProducto(Long id) {
+    public void eliminarProducto(Integer id) {
         productoRepository.deleteById(id);
     }
 }
